+++ lib/event_bloc.dart
@@ -7,6 +7,7 @@
 library event_bloc;
 
 export 'package:event_bloc/src/event_bloc/bloc.dart';
+export 'package:event_bloc/src/event_bloc/debugger.dart';
 export 'package:event_bloc/src/event_bloc/event.dart';
 export 'package:event_bloc/src/event_bloc/event_channel.dart';
 export 'package:event_bloc/src/event_bloc/refresh.dart';